Start add to cart works. Start show cart impl

+++ src/main/java/com/dudis/foodorders/domain/OrderItem.java
@@ -4,13 +4,15 @@ import lombok.Builder;
 import lombok.Value;
 import lombok.With;
 
+import java.math.BigDecimal;
+
 @With
 @Value
 @Builder
 public class OrderItem {
 
     Integer orderItemId;
-    Integer quantity;
+    BigDecimal quantity;
     Cart cart;
     Food food;
     Order order;
